ai-voice-agent/requirements.txt
elevenlabs==2.44.0
python-dotenv==1.2.2